+++ src/App.vue
@@ -20,7 +20,7 @@
 <script lang="ts">
 import { RouterView } from 'vue-router';
 import { defineComponent } from 'vue';
-import ErrorBoundary from '@/components/error/ErrorBoundary.vue';
+import ErrorBoundary from '@/components/ui/error/ErrorBoundary.vue';
 
 export default defineComponent({
   components: {
