+++ src/components/pages/course/lesson/SkillItem.vue
