+++ src/components/pages/course/lesson/VideoLayers/LockedLayer.vue
